sdk/go/ResourceOption: Provider option takes precedence
This change fixes a discrepancy in the Provider and Providers options
versus other SDKs:
Provider option doesn't always take precedence in the Go SDK.

## Background

Suppose we have providers 'ambient' and 'local' for the same package.
We want to use 'local' directly for the resource,
and 'ambient' for its child resources.
So we pass the options,
`Providers(ambient)` to put ambient into the bag of providers,
and `Provider(local)` to use it for the resource.

    NewResource(ctx, Providers(ambient), Provider(local))

This works as expected, but if we flip the order of the options,
it doesn't:

    NewResource(ctx, Provider(local), Providers(ambient))

This is because of how the two options are merged in the Go SDK.
`Provider(..)` unconditionally turns into `Providers(..)`,
losing the information about it being explicitly set for that resource.

https://github.com/pulumi/pulumi/blob/fbc66dd29e922eab3ce7870110ae55b5366cff1e/sdk/go/pulumi/resource.go#L458-L462

## Fix

Unfortunately, we can't just stop merging Provider into Providers.
In Go, we use functional options to implement resource options.
Unlike other SDKs, this makes it possible to provide the same option
multiple times.

Due to the prior merging logic, we've established in the Go SDK
that the following two are roughly equivalent:

    NewResource(.., Provider(p1), Provider(p2), Provider(p3))
    NewResource(.., Providers(p1, p2, p3))

If we change this now, it is likely to break users of the Go SDK.
Therefore, we need to retain this part of the behavior.

To fix this, we separately track the last passed `Provider`.
This will take precedence when we try to get the provider for a
resource.

Note:
We already had a Provider field on resourceOptions,
as well as machinery to have it take precedence.
We simply neglected to set that field when we refactored
ResourceOption merging in the past.

Resolves #12192