Set the current project as ambient context on backends.
Rather than passing a `*workspace.Project` to most backend methods, just
say that backends have a "current project" property. We set this on
creation (because we normally have a project before creating a backend)
but we also add a `SetCurrentProject` method for `pulumi new` to use
because it's the (currently) one place where we need a backend before we
have the project to go with it.